# feat(storage): add Bloom filter based event lookups