--- a/crates/pathfinder/src/bin/pathfinder/config.rs
+++ b/crates/pathfinder/src/bin/pathfinder/config.rs
@@ -224,6 +224,15 @@ This should only be enabled for debugging purposes as it adds substantial proces
         default_value = "500"
     )]
     get_events_max_blocks_to_scan: std::num::NonZeroUsize,
+
+    #[arg(
+        long = "rpc.get-events-max-bloom-filters-to-load",
+        long_help = "The number of Bloom filters to load for events when querying for events. \
+            This limit is used to prevent queries from taking too long.",
+        env = "PATHFINDER_RPC_GET_EVENTS_MAX_BLOOM_FILTERS_TO_LOAD",
+        default_value = "100000"
+    )]
+    get_events_max_bloom_filters_to_load: std::num::NonZeroUsize,
 }
 
 #[derive(clap::ValueEnum, Debug, Clone, Copy, PartialEq)]
@@ -482,6 +491,7 @@ pub struct Config {
     pub gateway_api_key: Option<String>,
     pub event_bloom_filter_cache_size: NonZeroUsize,
     pub get_events_max_blocks_to_scan: NonZeroUsize,
+    pub get_events_max_bloom_filters_to_load: NonZeroUsize,
 }
 
 pub struct Ethereum {
@@ -654,6 +664,7 @@ impl Config {
             gateway_api_key: cli.gateway_api_key,
             event_bloom_filter_cache_size: cli.event_bloom_filter_cache_size,
             get_events_max_blocks_to_scan: cli.get_events_max_blocks_to_scan,
+            get_events_max_bloom_filters_to_load: cli.get_events_max_bloom_filters_to_load,
         }
     }
 }