--- a/crates/pathfinder/src/bin/pathfinder/main.rs
+++ b/crates/pathfinder/src/bin/pathfinder/main.rs
@@ -173,7 +173,8 @@ Hint: This is usually caused by exceeding the file descriptor limit of your syst
     let rpc_config = pathfinder_rpc::context::RpcConfig {
         batch_concurrency_limit: config.rpc_batch_concurrency_limit,
         get_events_max_blocks_to_scan: config.get_events_max_blocks_to_scan,
-        get_events_max_bloom_filters_to_load: config.get_events_max_bloom_filters_to_load,
+        get_events_max_uncached_bloom_filters_to_load: config
+            .get_events_max_uncached_bloom_filters_to_load,
     };
 
     let context = pathfinder_rpc::context::RpcContext::new(