--- a/crates/pathfinder/src/state/sync.rs
+++ b/crates/pathfinder/src/state/sync.rs
@@ -671,9 +671,8 @@ async fn l1_update(
             .context("Insert update")?;
 
         let l2_hash = transaction
-            .block_id(update.block_number.into())
-            .context("Fetching block hash")?
-            .map(|(_, hash)| hash);
+            .block_hash(update.block_number.into())
+            .context("Fetching block hash")?;
 
         if let Some(l2_hash) = l2_hash {
             if l2_hash == update.block_hash {