--- a/crates/rpc/src/context.rs
+++ b/crates/rpc/src/context.rs
@@ -15,7 +15,7 @@ use tokio::sync::watch as tokio_watch;
 pub struct RpcConfig {
     pub batch_concurrency_limit: NonZeroUsize,
     pub get_events_max_blocks_to_scan: NonZeroUsize,
-    pub get_events_max_bloom_filters_to_load: NonZeroUsize,
+    pub get_events_max_uncached_bloom_filters_to_load: NonZeroUsize,
 }
 
 #[derive(Clone)]
@@ -83,7 +83,7 @@ impl RpcContext {
         let config = RpcConfig {
             batch_concurrency_limit: NonZeroUsize::new(8).unwrap(),
             get_events_max_blocks_to_scan: NonZeroUsize::new(1000).unwrap(),
-            get_events_max_bloom_filters_to_load: NonZeroUsize::new(1000).unwrap(),
+            get_events_max_uncached_bloom_filters_to_load: NonZeroUsize::new(1000).unwrap(),
         };
 
         Self::new(