--- a/crates/rpc/src/v03/method/get_events.rs
+++ b/crates/rpc/src/v03/method/get_events.rs
@@ -171,7 +171,7 @@ pub async fn get_events(
             .events(
                 &filter,
                 context.config.get_events_max_blocks_to_scan,
-                context.config.get_events_max_bloom_filters_to_load,
+                context.config.get_events_max_uncached_bloom_filters_to_load,
             )
             .map_err(|e| match e {
                 EventFilterError::PageSizeTooBig(_) => GetEventsError::PageSizeTooBig,