--- a/crates/storage/src/bloom.rs
+++ b/crates/storage/src/bloom.rs
@@ -1,6 +1,7 @@
 use bloomfilter::Bloom;
 use pathfinder_crypto::Felt;
 
+#[derive(Clone)]
 pub(crate) struct BloomFilter(Bloom<Felt>);
 
 impl BloomFilter {