This PR implements a new algorithm for serving `starknet_getEvents`.

---------------------------

### New Algorithm

Instead of relying on SQLite to evaluate the filter we're now simply iterating over the block range covered by the filter, fetch events for a block and then evaluate the filter in code.

To help with performance we're adding a per-block Bloom filter that contains event addresses and keys and use that to avoid scanning event data for blocks that are known not to contain matches. Bloom filters take 2 KiB per block and are stored compressed in the database. An LRU cache helps us avoid reloading recently used Bloom filters.

The algorithm also has configurable limits on:

- the number of Bloom filters loaded from the database,
- the number of blocks scanned for matching events.

Upon hitting either of those limits we return with the current result set and a continuation token that can be used to continue where we've left on a subsequent request. Both of these limits (along with the Bloom filter LRU cache size) are configurable. These limits effectively allow setting an upper limit on how long a single query might take.

### Storage

A Bloom filter with the parameters we've chosen takes 2 KiB per block. Compression brings that down to an average of 1430 bytes per block, so that's roughly 700 MiB for current mainnet in the database. On the other hand, this allows us to drop the `starknet_events` table (and related indexes) which takes up slightly more than 20% of our database right now.


#### Migration

The migration step takes considerable time. There are two major steps:

- Computing and storing the Bloom filters for events (using the `starknet_events` table): this step takes slightly less than half an hour with mainnet on my setup.
- Dropping the `starknet_events` table (and related stuff). Unfortunately this step required us to switch to `journal_mode=delete` for the migration, because otherwise the tables being dropped are copied to the WAL which is _very_ expensive for the ~180 GiB of data involved.

| Network | Size before | Size after | Migration duration | Logs |
| ---------- | ------------ | ---------- | --------------------- | ----- |
| sepolia-testnet | 804 MiB | 651 MiB | 1s | [sepolia-testnet.txt](https://github.com/eqlabs/pathfinder/files/14019599/sepolia-testnet.txt) |
| goerli-testnet | 213 GiB | 164 GiB | 10m 48s | [goerli-testnet.txt](https://github.com/eqlabs/pathfinder/files/14019605/goerli-testnet.txt) |
| mainnet | 790 GiB | 602 GiB | 59m 12s | [mainnet.txt](https://github.com/eqlabs/pathfinder/files/14019608/mainnet.txt) |


"Size after" is after a vacuum (not included in migration).